Move everything read/write lock to lock.py
