+++ fuel/utils/lock.py
@@ -325,3 +325,60 @@ def release_lock():
     if get_lock.lock_is_enabled and get_lock.n_lock == 0:
         get_lock.start_time = None
         get_lock.unlocker.unlock()
+
+
+def get_writelock(filename):
+    """Obtain a writelock on a file.
+
+    Only one write lock may be held at any given time.
+
+    Parameters
+    ----------
+    filename : str
+        Name of the file on which to obtain a writelock
+
+    """
+
+    # write lock expect locks to be on folder. Since we want a lock on a
+    # file, we will have to ask write lock for a folder with a different
+    # name from the file we want a lock on or else write lock will
+    # try to create a folder with the same name as the file
+    lock.get_lock(filename + ".writelock")
+
+
+def release_writelock():
+    """Release the previously obtained writelock."""
+    lock.release_lock()
+
+
+def release_readlock(lockdir_name):
+    """Release a previously obtained readlock
+
+    Parameters
+    ----------
+    lockdir_name : str
+        Name of the previously obtained readlock
+
+    """
+
+    # Make sure the lock still exists before deleting it
+    if os.path.exists(lockdir_name) and os.path.isdir(lockdir_name):
+        os.rmdir(lockdir_name)
+
+
+def get_readlock(pid, path):
+    """Obtain a readlock on a file
+
+    Parameters
+    ----------
+    path : str
+        Name of the file on which to obtain a readlock
+
+    """
+
+    timestamp = int(time.time() * 1e6)
+    lockdir_name = "%s.readlock.%i.%i" % (path, pid, timestamp)
+    os.mkdir(lockdir_name)
+
+    # Register function to release the readlock at the end of the script
+    atexit.register(release_readlock, lockdirName=lockdir_name)
